projeto-final-programacaoWeb/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
asyncpg
pydantic
passlib
python-jose
#pip install -r requirements.txt